\name{hpsCFA}
\alias{hpsCFA}
\title{Analisis Validitas Konstruk dengan CFA}
\usage{
hpsCFA()
}
\value{No return value, launches 'shiny' application}
\description{
Analisis CFA untuk melakukan validitas kontruk.Selain gratis, Package ini diharapkan dapat memberikan kemudahan dibandingkan menggunakan software-sofware berbayar.
}
\examples{
\dontrun{
library(shiny)
library(readxl)
library(psych)
library(corrplot)
library(lavaan)
library(semPlot)
library(writexl)
hpsCFA()
}
}

